+++ gems/aws-sdk-omics/lib/aws-sdk-omics.rb
@@ -55,7 +55,7 @@ module Aws::Omics
   autoload :EndpointProvider, 'aws-sdk-omics/endpoint_provider'
   autoload :Endpoints, 'aws-sdk-omics/endpoints'
 
-  GEM_VERSION = '1.63.0'
+  GEM_VERSION = '1.64.0'
 
 end